+++ src/core/base.jl
@@ -291,9 +291,9 @@ function instantiate_model_decomposition(
     multinetwork::Bool=false, pmitd_ref_extensions::Vector{<:Function}=Function[],
     export_models::Bool=false, kwargs...)
 
-    # Separate pmd ckts in a single dictionary to multiple dict entries
-    pmd_separated = _separate_pmd_circuits(pmitd_data["it"][_PMD.pmd_it_name]; multinetwork=multinetwork)
-    pmitd_data["it"][_PMD.pmd_it_name] = pmd_separated
+    # Separate distro. ckts from a single dictionary to multiple dictionary entries
+    distro_systems_separated = _separate_pmd_circuits(pmitd_data["it"][_PMD.pmd_it_name]; multinetwork=multinetwork)
+    pmitd_data["it"][_PMD.pmd_it_name] = distro_systems_separated
 
     # Correct the network data and assign the respective boundary number values.
     correct_network_data_decomposition!(pmitd_data; multinetwork=multinetwork)
@@ -303,71 +303,104 @@ function instantiate_model_decomposition(
 
     # ----- StsDOpt Optimizer ------
 
-    # PM models
-    pmitd_data["it"][_PM.pm_it_name][pmitd_it_name] = pmitd_data["it"][pmitd_it_name]         # add pmitd(boundary) info. to pm ref
+    # Add pmitd(boundary) info. to pm ref
+    pmitd_data["it"][_PM.pm_it_name][pmitd_it_name] = pmitd_data["it"][pmitd_it_name]
 
     # Instantiate the PM model
-    pm_inst_model = _IM.instantiate_model(pmitd_data["it"][_PM.pm_it_name],
+    master_instantiated = _IM.instantiate_model(pmitd_data["it"][_PM.pm_it_name],
                                     pmitd_type.parameters[1],
                                     build_method,
                                     ref_add_core_decomposition_transmission!,
                                     _PM._pm_global_keys,
-                                    _PM.pm_it_sym; kwargs...)
+                                    _PM.pm_it_sym; kwargs...
+    )
 
-    decomposed_models.pm = pm_inst_model
-    # Export nl models
+    # Add master model to struct
+    decomposed_models.pm = master_instantiated
+
+    # Export mof.json models
     if (export_models == true)
-        JuMP.write_to_file(pm_inst_model.model, "master_model_exported.mof.json")
+        JuMP.write_to_file(master_instantiated.model, "master_model_exported.mof.json")
     end
-    optimizer.master = pm_inst_model.model                                      # Add pm model to master
-    JuMP.set_optimizer(optimizer.master, _SDO.Optimizer; add_bridges = true)   # Set optimizer
 
-    # PMD models & Boundary linking vars
-    pmd_inst_models = []
-    pmd_inst_JuMP_models = []
-    boundary_vars_vect = []
+    # Add master model to optimizer master
+    optimizer.master = master_instantiated.model
 
+    # Set master optimizer
+    JuMP.set_optimizer(optimizer.master, _SDO.Optimizer; add_bridges = true)
+
+    # Get the number of subproblems
+    number_of_subproblems = length(pmitd_data["it"][_PMD.pmd_it_name])
+
+    # Convert distro. dictionary to vectors of dictionaries so it can be used in threaded version
+    ckts_names_vector = Vector{String}(undef, number_of_subproblems)
+    ckts_data_vector = Vector{Dict}(undef, number_of_subproblems)
+    ckt_number = 0
     for (ckt_name, ckt_data) in pmitd_data["it"][_PMD.pmd_it_name]
+        ckt_number = ckt_number + 1
+        ckts_names_vector[ckt_number] = ckt_name
+        ckts_data_vector[ckt_number] = ckt_data
+    end
+
+    # Set-up and instantiate subproblem models & boundary linking vars
+    subproblems_instantiated_models = Vector{pmitd_type.parameters[2]}(undef, number_of_subproblems)
+    subproblems_JuMP_models = Vector{JuMP.Model}(undef, number_of_subproblems)
+    boundary_vars_vector = Vector{Vector{Vector{JuMP.VariableRef}}}(undef, number_of_subproblems)
+
+    # Threaded loop for instantiating subproblems
+    Threads.@threads for i in 1:1:number_of_subproblems
 
         # Obtain ckt boundary data
         boundary_info = pmitd_data["it"][pmitd_it_name]
-        boundary_number = findfirst(x -> ckt_name == x["ckt_name"], boundary_info)
+        boundary_number = findfirst(x -> ckts_names_vector[i] == x["ckt_name"], boundary_info)
         boundary_for_ckt = Dict(boundary_number => boundary_info[boundary_number])
-        ckt_data[pmitd_it_name] = boundary_for_ckt                                  # add pmitd(boundary) info. to pmd ref
+
+        # add pmitd(boundary) info. to pmd ref
+        ckts_data_vector[i][pmitd_it_name] = boundary_for_ckt
 
         # Instantiate the PMD model
-        pmd_inst_model = _IM.instantiate_model(ckt_data,
+        subproblem_instantiated = _IM.instantiate_model(ckts_data_vector[i],
                                         pmitd_type.parameters[2],
                                         build_method,
                                         ref_add_core_decomposition_distribution!,
                                         _PMD._pmd_global_keys,
-                                        _PMD.pmd_it_sym; kwargs...)
+                                        _PMD.pmd_it_sym; kwargs...
+        )
 
-        push!(pmd_inst_models, pmd_inst_model)                                              # Add pmd IM model to vector
-        # Export nl models
+        # Add instantiated subproblem to vector of instantiated subproblems
+        subproblems_instantiated_models[i] = subproblem_instantiated
+
+        # Export mof.json models
         if (export_models == true)
-            JuMP.write_to_file(pmd_inst_model.model, "subproblem_$(ckt_name)_$(boundary_number)_model_exported.mof.json")
+            JuMP.write_to_file(subproblem_instantiated.model, "subproblem_$(i)_$(ckts_names_vector[i])_$(boundary_number)_model_exported.mof.json")
         end
-        JuMP.set_optimizer(pmd_inst_model.model , _SDO.Optimizer; add_bridges = true)      # Set the IDEC optimizer to the JuMP model
-        push!(pmd_inst_JuMP_models, pmd_inst_model.model )                                  # push the subproblem JuMP model into the vector of subproblems
 
-        # Boundary linking vars.
+        # Set the optimizer to the instantiated subproblem JuMP model
+        JuMP.set_optimizer(subproblem_instantiated.model, _SDO.Optimizer; add_bridges = true)
+
+        # Add the subproblem JuMP model into the vector of instantiated subproblems
+        subproblems_JuMP_models[i] = subproblem_instantiated.model
+
+        # Generate the boundary linking vars. (ACP, ACR, etc.)
         if (export_models == true)
-            linking_vars_vect = generate_boundary_linking_vars(pm_inst_model, pmd_inst_model, boundary_number; export_models=export_models)  # generates the respective (ACP, ACR, etc.) boundary linking vars vector.
+            linking_vars_vector = generate_boundary_linking_vars(master_instantiated, subproblem_instantiated, boundary_number; export_models=export_models)
         else
-            linking_vars_vect = generate_boundary_linking_vars(pm_inst_model, pmd_inst_model, boundary_number)  # generates the respective (ACP, ACR, etc.) boundary linking vars vector.
+            linking_vars_vector = generate_boundary_linking_vars(master_instantiated, subproblem_instantiated, boundary_number)
         end
 
-        push!(boundary_vars_vect, linking_vars_vect)                                                        # Add linking vars vector to vector containing all vectors of linking vars.
+        # Add linking vars vector to vector containing all vectors of linking vars.
+        boundary_vars_vector[i] = linking_vars_vector
 
     end
 
-    # Add subproblems
-    decomposed_models.pmd = pmd_inst_models          # Add all IM models to DecompositionStruct
-    optimizer.subproblems = pmd_inst_JuMP_models     # Add all pmd JuMP models (i.e., vector) as subproblems to Optimizer
+    # Add all instantiated subproblem models to DecompositionStruct
+    decomposed_models.pmd = subproblems_instantiated_models
 
-    # Boundary Linking vars
-    optimizer.list_linking_vars = boundary_vars_vect
+    # Add vector of subproblems JuMP models to Optimizer
+    optimizer.subproblems = subproblems_JuMP_models
+
+    # Add vecor of boundary linking vars to Optimizer
+    optimizer.list_linking_vars = boundary_vars_vector
 
     # Add Optimizer to DecompositionStruct
     decomposed_models.optimizer = optimizer